pyproject.toml
[tool.poetry]
name = "dbp_python_course"
version = "0.1.0"
description = ""
authors = []

[tool.black]
line-length = 100

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.12b0"
pytest-cov = "^3.0.0"
mypy = "^0.920"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"